+++ src/Luban.Core/BuiltinOptionNames.cs
@@ -84,6 +84,4 @@ public static class BuiltinOptionNames
     public const string IncrementalSidecarPath = "incremental.sidecarPath";
 
     public const string IncrementalOutputDir = "incremental.outputDir";
-
-    public const string IncrementalL10NChecksumPath = "incremental.l10nChecksumPath";
 }